Minor code review

+++ d3d11/d3d11.cpp
@@ -92,7 +92,7 @@ bool Test01()
     HRESULT hr = CreateDevice(device.GetAddressOf(), nullptr);
     if (FAILED(hr))
     {
-        printe("Failed creating device (HRESULT %08X)\n", hr);
+        printe("Failed creating device (HRESULT %08X)\n", static_cast<unsigned int>(hr));
         return false;
     }
 
@@ -134,7 +134,7 @@ bool Test01()
         if (FAILED(hr))
         {
             success = false;
-            printe("Failed getting data from (HRESULT %08X):\n%ls\n", hr, szPath);
+            printe("Failed getting data from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
         }
         else if (memcmp(&metadata, check, sizeof(TexMetadata)) != 0)
         {
@@ -165,7 +165,7 @@ bool Test01()
                 {
                     success = false;
                     pass = false;
-                    printe("Failed testing data from (HRESULT %08X):\n%ls\n", hr, szPath);
+                    printe("Failed testing data from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                 }
             }
 
@@ -191,7 +191,7 @@ bool Test02()
     HRESULT hr = CreateDevice(device.GetAddressOf(), nullptr);
     if (FAILED(hr))
     {
-        printe("Failed creating device (HRESULT %08X)\n", hr);
+        printe("Failed creating device (HRESULT %08X)\n", static_cast<unsigned int>(hr));
         return false;
     }
 
@@ -239,7 +239,7 @@ bool Test02()
         if (FAILED(hr))
         {
             success = false;
-            printe("Failed getting data from (HRESULT %08X):\n%ls\n", hr, szPath);
+            printe("Failed getting data from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
         }
         else if (memcmp(&metadata, check, sizeof(TexMetadata)) != 0)
         {
@@ -270,7 +270,7 @@ bool Test02()
             if (FAILED(hr))
             {
                 success = false;
-                printe("Failed creating texture from (HRESULT %08X):\n%ls\n", hr, szPath);
+                printe("Failed creating texture from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
             }
             else
             {
@@ -280,7 +280,7 @@ bool Test02()
                 if (FAILED(hr))
                 {
                     success = false;
-                    printe("Failed creating texture ex from (HRESULT %08X):\n%ls\n", hr, szPath);
+                    printe("Failed creating texture ex from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                 }
                 else
                 {
@@ -305,7 +305,7 @@ bool Test03()
     HRESULT hr = CreateDevice(device.GetAddressOf(), nullptr);
     if (FAILED(hr))
     {
-        printe("Failed creating device (HRESULT %08X)\n", hr);
+        printe("Failed creating device (HRESULT %08X)\n", static_cast<unsigned int>(hr));
         return false;
     }
 
@@ -354,7 +354,7 @@ bool Test03()
         if (FAILED(hr))
         {
             success = false;
-            printe("Failed getting data from (HRESULT %08X):\n%ls\n", hr, szPath);
+            printe("Failed getting data from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
         }
         else if (memcmp(&metadata, check, sizeof(TexMetadata)) != 0)
         {
@@ -370,7 +370,7 @@ bool Test03()
             if (FAILED(hr))
             {
                 success = false;
-                printe("Failed creating SRV from (HRESULT %08X):\n%ls\n", hr, szPath);
+                printe("Failed creating SRV from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
             }
             else
             {
@@ -379,7 +379,7 @@ bool Test03()
                 if (FAILED(hr))
                 {
                     success = false;
-                    printe("Failed creating SRV ex from (HRESULT %08X):\n%ls\n", hr, szPath);
+                    printe("Failed creating SRV ex from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                 }
                 else
                 {
@@ -404,7 +404,7 @@ bool Test04()
     HRESULT hr = SetupRenderTest(device.GetAddressOf(), nullptr);
     if (FAILED(hr))
     {
-        printe("Failed creating device (HRESULT %08X)\n", hr);
+        printe("Failed creating device (HRESULT %08X)\n", static_cast<unsigned int>(hr));
         return false;
     }
 
@@ -453,7 +453,7 @@ bool Test04()
         if (FAILED(hr))
         {
             success = false;
-            printe("Failed getting data from (HRESULT %08X):\n%ls\n", hr, szPath);
+            printe("Failed getting data from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
         }
         else if (memcmp(&metadata, check, sizeof(TexMetadata)) != 0)
         {
@@ -469,7 +469,7 @@ bool Test04()
             if (FAILED(hr))
             {
                 success = false;
-                printe("Failed creating SRV from (HRESULT %08X):\n%ls\n", hr, szPath);
+                printe("Failed creating SRV from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
             }
             else
             {
@@ -500,7 +500,7 @@ bool Test05()
     HRESULT hr = CreateDevice(device.GetAddressOf(), context.GetAddressOf());
     if (FAILED(hr))
     {
-        printe("Failed creating device (HRESULT %08X)\n", hr);
+        printe("Failed creating device (HRESULT %08X)\n", static_cast<unsigned int>(hr));
         return false;
     }
 
@@ -563,7 +563,7 @@ bool Test05()
         if (FAILED(hr))
         {
             success = false;
-            printe("Failed getting data from (HRESULT %08X):\n%ls\n", hr, szPath);
+            printe("Failed getting data from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
         }
         else if (memcmp(&metadata, check, sizeof(TexMetadata)) != 0)
         {
@@ -598,7 +598,7 @@ bool Test05()
                 if (FAILED(hr))
                 {
                     success = false;
-                    printe("Failed creating texture from (HRESULT %08X):\n%ls\n", hr, szPath);
+                    printe("Failed creating texture from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                 }
                 else
                 {
@@ -608,7 +608,7 @@ bool Test05()
                     if (FAILED(hr))
                     {
                         success = false;
-                        printe("Failed capturing texture from (HRESULT %08X):\n%ls\n", hr, szPath);
+                        printe("Failed capturing texture from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                     }
                     else
                     {
@@ -640,7 +640,7 @@ bool Test05()
                             {
                                 success = false;
                                 pass = false;
-                                printe("Failed writing DDS to (HRESULT %08X):\n%ls\n", hr, szDestPath);
+                                printe("Failed writing DDS to (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szDestPath);
                             }
 
                             if (!IsPlanar(metadata.format))
@@ -651,7 +651,7 @@ bool Test05()
                                 {
                                     success = false;
                                     pass = false;
-                                    printe("Failed comparing captured image (HRESULT %08X):\n%ls\n", hr, szPath);
+                                    printe("Failed comparing captured image (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                                 }
                                 else if (fabs(mse) > 0.000001f)
                                 {
@@ -675,7 +675,7 @@ bool Test05()
                 {
                     success = false;
                     pass = false;
-                    printe("Failed creating test staging texture (HRESULT %08X):\n%ls\n", hr, szPath);
+                    printe("Failed creating test staging texture (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                 }
                 else
                 {
@@ -685,7 +685,7 @@ bool Test05()
                     if (FAILED(hr))
                     {
                         success = false;
-                        printe("Failed capturing texture from staging texture (HRESULT %08X):\n%ls\n", hr, szPath);
+                        printe("Failed capturing texture from staging texture (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                     }
                     else
                     {
@@ -724,7 +724,7 @@ bool Test05()
                             {
                                 success = false;
                                 pass = false;
-                                printe("Failed writing DDS to (HRESULT %08X):\n%ls\n", hr, szDestPath2);
+                                printe("Failed writing DDS to (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szDestPath2);
                             }
 
                             if (!IsPlanar(metadata.format))
@@ -735,7 +735,7 @@ bool Test05()
                                 {
                                     success = false;
                                     pass = false;
-                                    printe("Failed comparing captured image (HRESULT %08X):\n%ls\n", hr, szPath);
+                                    printe("Failed comparing captured image (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                                 }
                                 else if (fabs(mse) > 0.000001f)
                                 {
