+++ hdr/hdr.cpp
@@ -98,7 +98,7 @@ bool Test01()
         if ( FAILED(hr) )
         {
             success = false;
-            printe( "Failed getting data from (HRESULT %08X):\n%ls\n", hr, szPath );
+            printe( "Failed getting data from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath );
         }
         else if ( memcmp( &metadata, check, sizeof(TexMetadata) ) != 0 )
         {
@@ -150,7 +150,7 @@ bool Test02()
         if ( FAILED(hr) )
         {
             success = false;
-            printe( "Failed getting raw file data from (HRESULT %08X):\n%ls\n", hr, szPath );
+            printe( "Failed getting raw file data from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath );
         }
         else
         {
@@ -162,7 +162,7 @@ bool Test02()
             if ( FAILED(hr) )
             {
                 success = false;
-                printe( "Failed loading hdr from memory (HRESULT %08X):\n%ls\n", hr, szPath );
+                printe( "Failed loading hdr from memory (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath );
             }
             else if ( memcmp( &metadata, check, sizeof(TexMetadata) ) != 0 )
             {
@@ -178,7 +178,7 @@ bool Test02()
                 if ( FAILED(hr) )
                 {
                     success = false;
-                    printe( "Failed computing MD5 checksum of image data (HRESULT %08X):\n%ls\n", hr, szPath );
+                    printe( "Failed computing MD5 checksum of image data (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath );
                 }
                 else if ( memcmp( digest, g_TestMedia[index].md5, 16 ) != 0 )
                 {
@@ -237,7 +237,7 @@ bool Test03()
         if ( FAILED(hr) )
         {
             success = false;
-            printe( "Failed loading hdr from memory (HRESULT %08X):\n%ls\n", hr, szPath );
+            printe( "Failed loading hdr from memory (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath );
         }
         else if ( memcmp( &metadata, check, sizeof(TexMetadata) ) != 0 )
         {
@@ -253,7 +253,7 @@ bool Test03()
             if ( FAILED(hr) )
             {
                 success = false;
-                printe( "Failed computing MD5 checksum of image data (HRESULT %08X):\n%ls\n", hr, szPath );
+                printe( "Failed computing MD5 checksum of image data (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath );
             }
             else if ( memcmp( digest, g_TestMedia[index].md5, 16 ) != 0 )
             {
@@ -308,7 +308,7 @@ bool Test04()
         if (FAILED(hr))
         {
             success = false;
-            printe("Failed loading HJDR from (HRESULT %08X):\n%ls\n", hr, szPath);
+            printe("Failed loading HJDR from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
         }
         else
         {
@@ -327,7 +327,7 @@ bool Test04()
             {
                 success = false;
                 pass = false;
-                printe("Failed writing hdr to memory (HRESULT %08X):\n%ls\n", hr, szPath);
+                printe("Failed writing hdr to memory (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
             }
             else
             {
@@ -338,7 +338,7 @@ bool Test04()
                 {
                     success = false;
                     pass = false;
-                    printe("Failed reading back written hdr to memory (HRESULT %08X):\n%ls\n", hr, szPath);
+                    printe("Failed reading back written hdr to memory (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                 }
                 else if (metadata.width != metadata2.width
                     || metadata.height != metadata2.height
@@ -363,7 +363,7 @@ bool Test04()
                     {
                         success = false;
                         pass = false;
-                        printe("Failed computing hdr memory readback MSE (HRESULT %08X):\n%ls\n", hr, szPath);
+                        printe("Failed computing hdr memory readback MSE (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                     }
                     else if (IsErrorTooLarge(mse, 0.001f))
                     {
@@ -393,7 +393,7 @@ bool Test04()
                 {
                     success = false;
                     pass = false;
-                    printe("Failed writing hdr to memory 96bpp (HRESULT %08X):\n%ls\n", hr, szPath);
+                    printe("Failed writing hdr to memory 96bpp (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                 }
                 else
                 {
@@ -404,7 +404,7 @@ bool Test04()
                     {
                         success = false;
                         pass = false;
-                        printe("Failed reading back written hdr to memory 96bpp (HRESULT %08X):\n%ls\n", hr, szPath);
+                        printe("Failed reading back written hdr to memory 96bpp (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                     }
                     else if (metadata.width != metadata2.width
                         || metadata.height != metadata2.height
@@ -429,7 +429,7 @@ bool Test04()
                         {
                             success = false;
                             pass = false;
-                            printe("Failed computing hdr memory readback MSE 96bpp (HRESULT %08X):\n%ls\n", hr, szPath);
+                            printe("Failed computing hdr memory readback MSE 96bpp (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                         }
                         else if (IsErrorTooLarge(mse, 0.001f))
                         {
@@ -510,7 +510,7 @@ bool Test05()
         if (FAILED(hr))
         {
             success = false;
-            printe("Failed loading JDR from (HRESULT %08X):\n%ls\n", hr, szPath);
+            printe("Failed loading JDR from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
         }
         else
         {
@@ -528,7 +528,7 @@ bool Test05()
             {
                 success = false;
                 pass = false;
-                printe("Failed writing hdr to (HRESULT %08X):\n%ls\n", hr, szDestPath);
+                printe("Failed writing hdr to (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szDestPath);
             }
             else
             {
@@ -539,7 +539,7 @@ bool Test05()
                 {
                     success = false;
                     pass = false;
-                    printe("Failed reading back written hdr to (HRESULT %08X):\n%ls\n", hr, szDestPath);
+                    printe("Failed reading back written hdr to (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szDestPath);
                 }
                 else if (metadata.width != metadata2.width
                     || metadata.height != metadata2.height
@@ -564,7 +564,7 @@ bool Test05()
                     {
                         success = false;
                         pass = false;
-                        printe("Failed computing hdr realoaded readback MSE (HRESULT %08X):\n%ls\n", hr, szDestPath);
+                        printe("Failed computing hdr realoaded readback MSE (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szDestPath);
                     }
                     else if (IsErrorTooLarge(mse, 0.001f))
                     {
@@ -593,7 +593,7 @@ bool Test05()
                 {
                     success = false;
                     pass = false;
-                    printe("Failed writing hdr to (HRESULT %08X):\n%ls\n", hr, szDestPath2);
+                    printe("Failed writing hdr to (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szDestPath2);
                 }
                 else
                 {
@@ -604,7 +604,7 @@ bool Test05()
                     {
                         success = false;
                         pass = false;
-                        printe("Failed reading back written hdr to (HRESULT %08X):\n%ls\n", hr, szDestPath2);
+                        printe("Failed reading back written hdr to (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szDestPath2);
                     }
                     else if (metadata.width != metadata2.width
                         || metadata.height != metadata2.height
@@ -629,7 +629,7 @@ bool Test05()
                         {
                             success = false;
                             pass = false;
-                            printe("Failed computing hdr realoaded readback MSE (HRESULT %08X):\n%ls\n", hr, szDestPath2);
+                            printe("Failed computing hdr realoaded readback MSE (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szDestPath2);
                         }
                         else if (IsErrorTooLarge(mse, 0.001f))
                         {
@@ -720,7 +720,7 @@ bool Test06()
                     if (FAILED(hr))
                     {
                         success = false;
-                        printe("Failed getting raw file data from (HRESULT %08X):\n%ls\n", hr, szPath);
+                        printe("Failed getting raw file data from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath);
                     }
                     else
                     {
@@ -733,7 +733,7 @@ bool Test06()
                             if (hr != HRESULT_FROM_WIN32(ERROR_NOT_SUPPORTED))
                             {
                                 success = false;
-                                printe("ERROR: frommemory expected success! (%08X)\n%ls\n", hr, szPath);
+                                printe("ERROR: frommemory expected success! (%08X)\n%ls\n", static_cast<unsigned int>(hr), szPath);
                             }
                         }
                         else if (SUCCEEDED(hr) && !ishdr)
@@ -756,7 +756,7 @@ bool Test06()
                     if (hr != HRESULT_FROM_WIN32(ERROR_NOT_SUPPORTED))
                     {
                         success = false;
-                        printe("ERROR: fromfile expected success ! (%08X)\n%ls\n", hr, szPath);
+                        printe("ERROR: fromfile expected success ! (%08X)\n%ls\n", static_cast<unsigned int>(hr), szPath);
                     }
                 }
                 else if (SUCCEEDED(hr) && !ishdr)
