+++ tex/fliprotate.cpp
@@ -141,7 +141,7 @@ bool TEXTest::Test04()
         if ( FAILED(hr) )
         {
             success = false;
-            printe( "Failed getting DDS data from (HRESULT %08X):\n%ls\n", hr, szPath );
+            printe( "Failed getting DDS data from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath );
         }
         else if ( memcmp( &metadata, check, sizeof(TexMetadata) ) != 0 )
         {
@@ -165,7 +165,7 @@ bool TEXTest::Test04()
             {
                 success = false;
                 pass = false;
-                printe( "Failed flip/rotating (HRESULT %08X) using %08X:\n%ls\n", hr, dwFlags, szPath );
+                printe( "Failed flip/rotating (HRESULT %08X) using %08X:\n%ls\n", static_cast<unsigned int>(hr), dwFlags, szPath );
             }
             else
             {
@@ -226,7 +226,7 @@ bool TEXTest::Test04()
                 {
                     success = false;
                     pass = false;
-                    printe( "Failed flip/rotating complex (HRESULT %08X) using %08X:\n%ls\n", hr, dwFlags, szPath );
+                    printe( "Failed flip/rotating complex (HRESULT %08X) using %08X:\n%ls\n", static_cast<unsigned int>(hr), dwFlags, szPath );
                 }
                 else
                 {
