+++ tex/planar.cpp
@@ -83,7 +83,7 @@ bool TEXTest::Test15()
         if ( FAILED(hr) )
         {
             success = false;
-            printe( "Failed getting DDS data from (HRESULT %08X):\n%ls\n", hr, szPath );
+            printe( "Failed getting DDS data from (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath );
         }
         else if ( memcmp( &metadata, check, sizeof(TexMetadata) ) != 0 )
         {
@@ -103,7 +103,7 @@ bool TEXTest::Test15()
             {
                 success = false;
                 pass = false;
-                printe( "Failed single plane conversion (HRESULT %08X):\n%ls\n", hr, szPath );
+                printe( "Failed single plane conversion (HRESULT %08X):\n%ls\n", static_cast<unsigned int>(hr), szPath );
             }
             else if ( image.GetMetadata().format != g_TestMedia[index].tformat
                       || image.GetMetadata().width != srcimage.GetMetadata().width || image.GetMetadata().height != srcimage.GetMetadata().height )
